client/requirements.txt
requests==2.31.0
python-socketio==5.8.0
